@@ -12,31 +12,32 @@
   "author": "Igor Korsakov <overtorment@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/cli": "^7.0.0",
-    "babel-eslint": "^10.1.0",
-    "@babel/node": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/register": "^7.0.0",
+    "@babel/cli": "^7.12.10",
+    "@babel/core": "^7.12.10",
+    "@babel/eslint-parser": "^7.12.1",
+    "@babel/node": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/register": "^7.12.10",
+    "@grpc/grpc-js": "^1.2.2",
+    "@grpc/proto-loader": "^0.5.5",
     "bignumber.js": "^9.0.1",
-    "bolt11": "^1.2.6",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "express": "^4.16.4",
+    "bolt11": "^1.2.7",
+    "core-js": "^3.8.1",
+    "eslint": "^7.16.0",
+    "eslint-config-prettier": "^7.1.0",
+    "eslint-plugin-prettier": "^3.3.0",
+    "express": "^4.17.1",
     "express-rate-limit": "^5.2.3",
     "frisbee": "^3.1.4",
-    "grpc": "^1.24.4",
-    "ioredis": "^4.19.2",
-    "jayson": "^3.3.4",
-    "morgan": "^1.9.1",
-    "mustache": "^4.0.1",
-    "node-uuid": "^1.4.8",
-    "prettier": "^2.2.0",
+    "ioredis": "^4.19.4",
+    "jayson": "^3.4.4",
+    "morgan": "^1.10.0",
+    "mustache": "^4.1.0",
+    "prettier": "^2.2.1",
     "qr-image": "3.2.0",
-    "request": "^2.88.0",
+    "request": "^2.88.2",
     "request-promise": "^4.2.6",
+    "uuid": "^8.3.2",
     "winston": "^3.3.3"
   }
 }
