@@ -1,4 +1,4 @@
-process.on('uncaughtException', function(err) {
+process.on('uncaughtException', function (err) {
   console.error(err);
   console.log('Node NOT Exiting...');
 });
@@ -6,7 +6,7 @@ process.on('uncaughtException', function(err) {
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
 let express = require('express');
 let morgan = require('morgan');
-let uuid = require('node-uuid');
+import { v4 as uuidv4 } from 'uuid';
 let logger = require('./utils/logger');
 
 morgan.token('id', function getId(req) {
@@ -23,8 +23,8 @@ const limiter = rateLimit({
 });
 app.use(limiter);
 
-app.use(function(req, res, next) {
-  req.id = uuid.v4();
+app.use(function (req, res, next) {
+  req.id = uuidv4();
   next();
 });
 
@@ -44,7 +44,7 @@ app.use('/static', express.static('static'));
 app.use(require('./controllers/api'));
 app.use(require('./controllers/website'));
 
-let server = app.listen(process.env.PORT || 3000, function() {
+let server = app.listen(process.env.PORT || 3000, function () {
   logger.log('BOOTING UP', 'Listening on port ' + (process.env.PORT || 3000));
 });
 module.exports = server;
