@@ -11,14 +11,14 @@ let lightningListChannels = {};
 function updateLightning() {
   console.log('updateLightning()');
   try {
-    lightning.getInfo({}, function(err, info) {
+    lightning.getInfo({}, function (err, info) {
       if (err) {
         console.error('lnd failure:', err);
       }
       lightningGetInfo = info;
     });
 
-    lightning.listChannels({}, function(err, response) {
+    lightning.listChannels({}, function (err, response) {
       if (err) {
         console.error('lnd failure:', err);
         return;
@@ -79,7 +79,7 @@ const pubkey2name = {
   '037cc5f9f1da20ac0d60e83989729a204a33cc2d8e80438969fadf35c1c5f1233b': 'lnd2.bluewallet.io',
 };
 
-router.get('/', function(req, res) {
+router.get('/', function (req, res) {
   logger.log('/', [req.id]);
   if (!lightningGetInfo) {
     console.error('lnd failure');
@@ -104,7 +104,7 @@ router.get('/about', function(req, res) {
   return res.status(200).send(mustache.render(html, {}));
 });
 
-router.use(function(req, res) {
+router.use(function (req, res) {
   res.status(404).send('404');
 });
 
